Update SDL2 for enum change in nim/devel.
Enums with holes are no longer ordinals, so attempts to put sdl2.EventType
into a set no longer compiles.

The only thing the sets were used for are kind assertions in the
``evConv`` template, for sets of one two six members. I replaced
the sets in the calls to evConv with arrays.

As added bonus, this uses less memory than a set[int32].